cleanup: move paginateData to its own file
Separates the pure utility from the hook file for cleaner module boundaries.
