+++ packages/core/src/Table/plugins/pagination/index.ts
@@ -1,2 +1,3 @@
-export {useXDSTablePagination, paginateData} from './useXDSTablePagination';
+export {useXDSTablePagination} from './useXDSTablePagination';
 export type {UseXDSTablePaginationConfig} from './useXDSTablePagination';
+export {paginateData} from './paginateData';
